Replace web build with standalone bundle and remove Python
