--- settings.py
@@ -1,7 +1,0 @@
-# settings.py
-
-# Screen dimensions
-WIDTH = 600
-HEIGHT = 880
-
-FULLSCREEN = False
