+++ web/index.html
@@ -11,6 +11,6 @@
       <canvas id="game-canvas" width="600" height="880"></canvas>
       <div id="overlay"></div>
     </div>
-    <script type="module" src="./js/main.js"></script>
+    <script src="./js/game.js"></script>
   </body>
 </html>
